# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API RotaVan
  description: API para gestão de transporte escolar, incluindo cadastro de responsáveis, motoristas, escolas, rotas, veículos, contratos e viagens.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento Local
...
security:
  - BearerAuth: [] # Aplica segurança Bearer a todos os endpoints por padrão (exceto os explicitamente excluídos)
paths:
...
  /v1/api/rotas:
    get:
      tags:
        - Rotas
      summary: Lista todas as rotas.
      operationId: getAllRotas
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso - Retorna a lista de rotas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rota'
        '401':
          description: Não autorizado.
    post:
      tags:
        - Rotas
      summary: Cria uma nova rota.
      operationId: createRota
      security:
        - BearerAuth: []
      requestBody:
        description: Objeto da rota a ser criada. Pontos podem ser incluídos opcionalmente.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotaInput' # Usando um schema de input
      responses:
        '201':
          description: Rota criada com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rota'
        '400':
          description: Requisição inválida.
        '401':
          description: Não autorizado.
  /v1/api/rotas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID da rota
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Rotas
      summary: Busca uma rota pelo ID.
      operationId: getRotaById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rota'
        '404':
          description: Rota não encontrada.
        '401':
          description: Não autorizado.
    put:
      tags:
        - Rotas
      summary: Atualiza uma rota existente.
      operationId: updateRota
      security:
        - BearerAuth: []
      requestBody:
        description: Objeto da rota com os dados atualizados.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotaInput' # Usando input schema
      responses:
        '200':
          description: Atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rota'
        '404':
          description: Rota não encontrada.
        '401':
          description: Não autorizado.
    delete:
      tags:
        - Rotas
      summary: Apaga uma rota (e seus pontos em cascata).
      operationId: deleteRota
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Apagada com sucesso (Sem conteúdo).
        '404':
          description: Rota não encontrada.
        '401':
          description: Não autorizado.
  /v1/api/rotas/{rotaId}/pontos:
    parameters:
      - name: rotaId
        in: path
        required: true
        description: ID da Rota à qual adicionar o ponto
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Rotas
      summary: Lista todos os pontos de uma rota específica, ordenados pela ordem.
      operationId: getPontosByRotaId
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso - Retorna a lista de pontos da rota.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ponto'
        '404':
          description: Rota não encontrada.
        '401':
          description: Não autorizado.
    post:
      tags:
        - Rotas
      summary: Adiciona um novo ponto a uma rota existente.
      operationId: addPontoToRota
      security:
        - BearerAuth: []
      requestBody:
        description: Objeto do ponto a ser adicionado (sem o ID da rota).
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PontoInput' # Usar schema de input
      responses:
        '201':
          description: Ponto adicionado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ponto'
        '400':
          description: Requisição inválida.
        '404':
          description: Rota não encontrada.
        '401':
          description: Não autorizado.

  # ===============================================================
  # ENDPOINTS DE PONTOS (OPCIONAL - NOVOS)
  # ===============================================================
  /v1/api/pontos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID do ponto
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Pontos
      summary: Busca um ponto pelo ID.
      operationId: getPontoById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ponto'
        '404':
          description: Ponto não encontrado.
        '401':
          description: Não autorizado.
    put:
      tags:
        - Pontos
      summary: Atualiza um ponto existente (sem alterar a rota).
      operationId: updatePonto
      security:
        - BearerAuth: []
      requestBody:
        description: Objeto do ponto com dados atualizados.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PontoInput' # Usar schema de input
      responses:
        '200':
          description: Atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ponto'
        '400':
          description: Requisição inválida.
        '404':
          description: Ponto não encontrado.
        '401':
          description: Não autorizado.
    delete:
      tags:
        - Pontos
      summary: Apaga um ponto pelo ID.
      operationId: deletePonto
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Apagado com sucesso.
        '404':
          description: Ponto não encontrado.
        '401':
          description: Não autorizado.

  # ===============================================================
  # ENDPOINTS DE VEICULOS (NOVOS)
  # ===============================================================
...
  /v1/api/contratos:
    get:
      tags:
        - Contratos
      summary: Lista todos os contratos.
      operationId: getAllContratos
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contrato'
        '401':
          description: Não autorizado.
    post:
      tags:
        - Contratos
      summary: Cria um novo contrato.
      operationId: createContrato
      security:
        - BearerAuth: []
      requestBody:
        description: Dados do contrato a ser criado.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratoInput'
      responses:
        '201':
          description: Contrato criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contrato'
        '400':
          description:
        '404':
          description: Responsável ou Motorista não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/contratos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID do contrato
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Contratos
      summary: Busca um contrato pelo ID.
      operationId: getContratoById
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contrato'
        '404':
          description: Contrato não encontrado.
        '401':
          description: Não autorizado.
    put:
      tags:
        - Contratos
      summary: Atualiza um contrato existente.
      operationId: updateContrato
      security:
        - BearerAuth: []
      requestBody:
        description: Dados do contrato atualizados.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContratoInput' # Input pode omitir FKs se não forem alteráveis
      responses:
        '200':
          description: Atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contrato'
        '400':
          description: Requisição inválida.
        '404':
          description: Contrato não encontrado.
        '401':
          description: Não autorizado.
    delete:
      tags:
        - Contratos
      summary: Apaga um contrato.
      operationId: deleteContrato
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Apagado com sucesso.
        '404':
          description: Contrato não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/contratos/responsavel/{responsavelId}:
    parameters:
      - name: responsavelId
        in: path
        required: true
        description: ID do Responsável
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Contratos
      summary: Lista contratos por ID do responsável.
      operationId: getContratosByResponsavel
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contrato'
        '404':
          description: Responsável não encontrado.
        '401':
          description: Não autorizado.
  /v1/api/contratos/motorista/{motoristaId}:
    parameters:
      - name: motoristaId
        in: path
        required: true
        description: ID do Motorista
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Contratos
      summary: Lista contratos por ID do motorista.
      operationId: getContratosByMotorista
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Sucesso.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contrato'
        '404':
          description: Motorista não encontrado.
        '401':
          description: Não autorizado.

  # ===============================================================
  # ENDPOINTS DE VIAGENS (NOVOS)
  # ===============================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Rota:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        nomeRota: { type: string }
        descricao: { type: string, nullable: true }
        distanciaKm: { type: number, format: double, nullable: true }
        tempoEstimadoMin: { type: integer, nullable: true }
        pontos:
          type: array
          items:
            $ref: '#/components/schemas/PontoInfo' # Simplificado para evitar recursão
          readOnly: true
          nullable: true
        # viagens: # Omitido para simplicidade na resposta principal da Rota
    RotaInput: # Para POST e PUT /rotas
      type: object
      required: [ nomeRota ]
      properties:
        nomeRota: { type: string }
        descricao: { type: string, nullable: true }
        distanciaKm: { type: number, format: double, nullable: true }
        tempoEstimadoMin: { type: integer, nullable: true }
        # Não inclui pontos ou viagens

    # Ponto
    Ponto:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        longitude: { type: number, format: double } # BigDecimal mapeado para number
        latitude: { type: number, format: double }
        nomePonto: { type: string, nullable: true }
        ordem: { type: integer }
        rota: { $ref: '#/components/schemas/RotaInfo' } # Simplificado
    PontoInput: # Para POST /rotas/{rotaId}/pontos e PUT /pontos/{id}
      type: object
      required: [ longitude, latitude, ordem ]
      properties:
        longitude: { type: number, format: double }
        latitude: { type: number, format: double }
        nomePonto: { type: string, nullable: true }
        ordem: { type: integer }
        # Não inclui Rota (vem do path ou não é alterado no PUT individual)

    # Veiculo
...
    Contrato:
      type: object
      properties:
        id: { type: string, format: uuid, readOnly: true }
        dataInicio: { type: string, format: date }
        dataFim: { type: string, format: date }
        valorMensal: { type: number, format: double } # BigDecimal
        responsavel: { $ref: '#/components/schemas/ResponsavelInfo' } # Simplificado
        motorista: { $ref: '#/components/schemas/MotoristaInfo' } # Simplificado
    ContratoInput: # Para POST e PUT /contratos
      type: object
      required: [ dataInicio, dataFim, valorMensal, responsavelId, motoristaId ]
      properties:
        dataInicio: { type: string, format: date }
        dataFim: { type: string, format: date }
        valorMensal: { type: number, format: double }
        responsavelId: { type: string, format: uuid, description: ID do responsável existente }
        motoristaId: { type: string, format: uuid, description: ID do motorista existente }

    # Viagem
...
    RotaInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomeRota: { type: string }
    PontoInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomePonto: { type: string, nullable: true }
        ordem: { type: integer }
    MotoristaInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomeMotorista: { type: string }
...
    ResponsavelInfo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nomeResponsavel: { type: string }
...